test(bindings): one conformance suite — the vocabulary-accounting fixture
The identical describe block triplicated across the targets moves to
tests/fixtures/ in the bindings package (the contract's home); each target
invokes it with its own normalize + ledgers.
